Refactor Onboardly component by removing the legacy Onboardly.tsx file and updating exports in index.ts; streamline README by removing outdated development instructions.
