+++ packages/onboardly/src/components/index.ts
@@ -1,0 +1,3 @@
+export { Onboardly } from './Onboardly';
+export { useOnboardly } from './context';
+export type { OnboardlyProps, OnboardlyStep } from './types';
